src/requirements.txt
colorama==0.4.6
exceptiongroup==1.2.1
iniconfig==2.0.0
loguru==0.7.2
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
PyYAML==6.0.1
tomli==2.0.1
win32-setctime==1.1.0
